# openapi.yaml
openapi: 3.1.0
info:
  title: HighFive! API
  version: 2.0.0
  description: >-
    Contrat backend v2, derive de `src/domain/**` (schemas zod), `src/api/*.ts`
    (appels reels du front) et `src/mocks/handlers/**` (regles metier
    appliquees cote mock). Voir `docs/v2/backend/SPEC.md` pour le detail des
    regles metier, du modele de persistance propose et des ecarts connus.
    Les schemas de composants sont generes depuis `src/domain/index.ts` par
    `pnpm export:schemas` (zod 4, `z.toJSONSchema`) dans `docs/v2/backend/schemas/`.
servers:
  - url: /api
    description: Prefixe fixe (voir `src/api/client.ts`), au-dessus de `VITE_API_URL`.
security:
  - bearerAuth: []
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Session courante
      responses:
        "200":
          description: Personne connectee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateCurrentUser
      tags: [users]
      summary: Mettre a jour son profil
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserProfileUpdateInput" }
      responses:
        "200":
          description: Profil mis a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CurrentUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: R-P1 — compte `suspended` (bio/avatar verrouilles).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
...
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Profil public
      security: []
      parameters:
        - $ref: "#/components/parameters/UsernameParam"
      responses:
        "200":
          description: Profil public.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "404":
          description: Introuvable ou compte `deleted`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /users/{username}/projects:
    get:
      operationId: getUserProjects
      tags: [users]
      summary: Projets d'une personne (R-V4)
      security: []
      parameters:
        - $ref: "#/components/parameters/UsernameParam"
      responses:
        "200":
          description: Projets crees, collaborations, et nombre de projets prives sans les nommer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProjectsResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /me/avatar:
    post:
      operationId: uploadAvatar
      tags: [files]
      summary: Deposer un nouvel avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar: { type: string, format: binary }
              required: [avatar]
      responses:
        "200":
          description: Nouvelle URL d'avatar.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvatarResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ------------------------------------------------------------- Projects
...
  /projects/{slug}/columns:
    get:
      operationId: listColumns
      tags: [tasks]
      summary: Colonnes des Taches
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Colonnes ordonnees.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Column" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: createColumn
      tags: [tasks]
      summary: Creer une colonne (R-K2 — 1 a 6)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ColumnCreateInput" }
      responses:
        "201":
          description: Colonne creee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Column" }
        "400":
          description: "Corps invalide, ou plafond de 6 colonnes atteint (R-K2)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /columns/{columnId}:
    delete:
      operationId: deleteColumn
      tags: [tasks]
      summary: Supprimer une colonne (R-K3 — destination obligatoire)
      parameters:
        - $ref: "#/components/parameters/ColumnIdParam"
        - name: moveTo
          in: query
          required: true
          schema: { $ref: "#/components/schemas/Id" }
          description: Colonne de destination pour les taches deplacees (R-K3).
      responses:
        "204": { description: Colonne supprimee, taches deplacees. }
        "400":
          description: "`moveTo` manquant, ou derniere colonne du projet (au moins 1 restante)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Colonne source ou colonne de destination introuvable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /projects/{slug}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: Taches d'un projet
      security: []
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: Taches.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Creer une tache (R-K4 — titre seul obligatoire)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskCreateInput" }
      responses:
        "201":
          description: Tache creee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Colonne cible introuvable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }
  /tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Modifier une tache
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskUpdateInput" }
      responses:
        "200":
          description: Tache mise a jour.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Supprimer une tache
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      responses:
        "204": { description: Tache supprimee. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /tasks/{taskId}/move:
    post:
      operationId: moveTask
      tags: [tasks]
      summary: Deplacer une tache entre colonnes (glisser-deposer)
      parameters:
        - $ref: "#/components/parameters/TaskIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskMoveInput" }
      responses:
        "200":
          description: Tache deplacee.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Tache ou colonne de destination introuvable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }

  # ---------------------------------------------------------------- Wall
  /projects/{slug}/wall:
    get:
      operationId: getWall
      tags: [wall]
      summary: Metadonnees du Mur (R-W1 — membre+/observateur)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      responses:
        "200":
          description: "Metadonnees (le document collaboratif tldraw/Yjs lui-meme ne passe pas par REST, voir SPEC.md)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Wall" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{slug}/wall/to-tasks:
    post:
      operationId: convertWallSelectionToTasks
      tags: [wall]
      summary: Convertir une selection du Mur en taches (R-W2)
      parameters:
        - $ref: "#/components/parameters/SlugParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WallToTasksInput" }
      responses:
        "201":
          description: 'Taches creees dans la colonne "A faire", lien vers l''origine conserve.'
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Projet introuvable, ou aucune colonne disponible.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiErrorBody" }

  # -------------------------------------------------------------- Files
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Jeton opaque (pas necessairement un JWT dans le mock : `nextId()`).
        Envoye `Authorization: Bearer <token>` par `src/api/client.ts`.
  parameters:
...
    SlugParam:
      name: slug
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Slug" }
    UsernameParam:
      name: username
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Username" }
...
    ColumnIdParam:
      name: columnId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
    TaskIdParam:
      name: taskId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Id" }
...
  responses:
    BadRequest:
      description: "`validation_error` — corps ou parametres invalides."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Unauthorized:
      description: "`unauthorized` — jeton absent ou invalide."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    Forbidden:
      description: "`forbidden` — role insuffisant pour cette action."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
    NotFound:
      description: "`not_found` — ressource introuvable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorBody" }
...
  schemas:
...
    ApiErrorBody: { $ref: "./schemas/ApiErrorBody.json" }
    Column: { $ref: "./schemas/Column.json" }
    ColumnCreateInput: { $ref: "./schemas/ColumnCreateInput.json" }
...
    CurrentUser: { $ref: "./schemas/CurrentUser.json" }
...
    Id: { $ref: "./schemas/Id.json" }
...
    ProjectSummary: { $ref: "./schemas/ProjectSummary.json" }
...
    Slug: { $ref: "./schemas/Slug.json" }
...
    Task: { $ref: "./schemas/Task.json" }
    TaskCreateInput: { $ref: "./schemas/TaskCreateInput.json" }
    TaskMoveInput: { $ref: "./schemas/TaskMoveInput.json" }
    TaskUpdateInput: { $ref: "./schemas/TaskUpdateInput.json" }
...
    User: { $ref: "./schemas/User.json" }
    UserProfileUpdateInput: { $ref: "./schemas/UserProfileUpdateInput.json" }
...
    Username: { $ref: "./schemas/Username.json" }
...
    Wall: { $ref: "./schemas/Wall.json" }
    WallToTasksInput: { $ref: "./schemas/WallToTasksInput.json" }
...
    UserProjectsResponse:
      description: "R-V4 (`userProjectsResponseSchema`, src/api/users.ts)."
      type: object
      properties:
        created:
          {
            type: array,
            items: { $ref: "#/components/schemas/ProjectSummary" },
          }
        collaborations:
          {
            type: array,
            items: { $ref: "#/components/schemas/ProjectSummary" },
          }
        privateProjectsCount: { type: integer, minimum: 0 }
      required: [created, collaborations, privateProjectsCount]
      additionalProperties: false
...
    AvatarResponse:
      description: "Reponse de `POST /me/avatar` (schema anonyme, src/api/files.ts)."
      type: object
      properties:
        avatar: { type: string, format: uri }
      required: [avatar]
      additionalProperties: false
...